Assumed that -1UL would work on all platforms.
    On Windows it used -1ULL.
    Modified the code to static_cast<std::size_t>(-1) this works on Windows.
    Will test on Linux and Mac.

+++ src/ThorsIOUtil/Formatter.h
@@ -16,6 +16,7 @@
 #include <typeindex>
 #include <cassert>
 #include <type_traits>
+#include <cstdint>
 
 
 namespace ThorsAnvil
@@ -366,7 +367,7 @@ class Formatter
                                             : info.width;
                 std::size_t fractPrec = (info.useDynamicSize == Dynamic::Precision || info.useDynamicSize == Dynamic::Both)
                                             ? s.iword(static_cast<int>(Dynamic::Precision))
-                                            : info.precision == -1UL && info.type == Type::Float ? 6 : info.precision;
+                                            : info.precision == static_cast<std::size_t>(-1) && info.type == Type::Float ? 6 : info.precision;
                 bool                    forceLeft = info.leftJustify;
                 std::ios_base::fmtflags format    = info.format;
                 if ((info.useDynamicSize == Dynamic::Width || info.useDynamicSize == Dynamic::Both) && s.iword(static_cast<int>(Dynamic::Width)) < 0)
